Refactoring voucher, reservation done, without reservation tests

+++ Backend/src/main/java/Reservista/example/Backend/Services/Registration/OTPService.java
@@ -3,6 +3,7 @@ package Reservista.example.Backend.Services.Registration;
 import Reservista.example.Backend.DAOs.OTPRepository;
 import Reservista.example.Backend.DAOs.UserRepository;
 import Reservista.example.Backend.Enums.StatusCode;
+import Reservista.example.Backend.Error.GlobalException;
 import Reservista.example.Backend.MailComponent.Mail;
 import Reservista.example.Backend.MailComponent.MailService;
 import Reservista.example.Backend.MailComponent.mailParsers.AccountActivationMailParser;
@@ -12,10 +13,9 @@ import Reservista.example.Backend.Models.EntityClasses.OTP;
 import Reservista.example.Backend.Models.EntityClasses.User;
 
 import Reservista.example.Backend.DTOs.Response.ResponseDTO;
-import Reservista.example.Backend.Models.EntityClasses.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import Reservista.example.Backend.Models.EntityClasses.OTP;
 
 import java.time.LocalDateTime;
 @Service
@@ -28,41 +28,38 @@ public class OTPService {
     @Autowired
     private UserRepository userRepository;
 
-    public ResponseDTO verifyGmailAccount(String email, String code) {
+    public void verifyGmailAccount(String email, String code) throws GlobalException {
 
-        ResponseDTO verifiedOTP = verifyOTP(email, code);
-        if (verifiedOTP.getStatus() != StatusCode.SUCCESS.getCode()) {
-            return verifiedOTP;
-        }
+        verifyOTP(email, code);
         User user = userRepository.findByEmail(email).orElse(null);
-        if (user == null || user.isActivated()) {
-            return StatusCode.INVALID_REQUEST.getRespond();
-        }
+        if (user == null || user.isActivated())
+            throw new GlobalException(StatusCode.INVALID_OTP_REQUEST, HttpStatus.BAD_REQUEST);
+
         user.setActivated(true);
         userRepository.save(user);
         Mail registrationMail = new RegistrationMailParser(user);
-        mailService.sendMail(registrationMail);
-        return StatusCode.SUCCESS.getRespond();
+        if (!mailService.sendMail(registrationMail)) throw new GlobalException(StatusCode.EMAIL_NOT_REACHED, HttpStatus.SERVICE_UNAVAILABLE); // 503
 
     }
 
-    public ResponseDTO refreshOTP(String email) {
+    public void refreshOTP(String email) throws GlobalException {
 
         User user = userRepository.findByEmail(email).orElse(null);
-        if (user == null) {
-            return StatusCode.INVALID_REQUEST.getRespond();
-        }
-        if (user.isActivated()) {
-            return StatusCode.INVALID_REQUEST.getRespond();
-        }
+        if (user == null || user.isActivated())
+            throw new GlobalException(StatusCode.INVALID_OTP_REQUEST, HttpStatus.BAD_REQUEST);
+
+
         OTP otp = otpRepository.findByEmail(email);
-        if (otp != null) {
+        if (otp != null)
             otpRepository.delete(otp);
-        }
-        return createAndSendOTP(user);
+
+        if (!createAndSendOTP(user))
+            throw new GlobalException(StatusCode.EMAIL_NOT_REACHED, HttpStatus.SERVICE_UNAVAILABLE); // 503
+
+
     }
 
-    public ResponseDTO createAndSendOTP(User user) {
+    public boolean createAndSendOTP(User user) {
 
         OTP newOtp = new OTP(user.getEmail());
         otpRepository.save(newOtp);
@@ -72,24 +69,26 @@ public class OTPService {
 
     }
 
-    public ResponseDTO verifyOTP(String email, String code) {
+    public boolean verifyOTP(String email, String code) throws GlobalException {
         OTP otp = otpRepository.findByEmail(email);
         if (otp == null) {
-            return StatusCode.NOT_REGISTERED_USER.getRespond();
+            throw new GlobalException(StatusCode.NOT_REGISTERED_USER, HttpStatus.NOT_FOUND );  ////// 404
         }
 
         if (!otp.getCode().equals(code)) {
             System.out.println("DB code: " + otp.getCode());
             System.out.println("sent code: " + code);
-            return StatusCode.WRONG_VERIFICATION_CODE.getRespond();
+            throw new GlobalException(StatusCode.WRONG_VERIFICATION_CODE, HttpStatus.UNPROCESSABLE_ENTITY ); //422 validation failure
         }
         LocalDateTime now=LocalDateTime.now();
         if (now.compareTo(otp.getExpirationDate()) > 0) { //is the time now after expiration date
             otpRepository.delete(otp);
-            return StatusCode.EXPIRED_VERIFICATION_COD.getRespond();
+
+            throw new GlobalException(StatusCode.EXPIRED_VERIFICATION_CODE, HttpStatus.GONE ); //410
         }
         otpRepository.delete(otp);
-        return StatusCode.SUCCESS.getRespond();
+        return true;
+
     }
